Move copyText to new file
